% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsim_plot_interactive.R
\name{rsim_plot_interactive}
\alias{rsim_plot_interactive}
\title{Interactive plot for rsim}
\usage{
rsim_plot_interactive(
  rsim_output,
  spname = "all",
  indplot = FALSE,
  palette = "rainbow"
)
}
\arguments{
\item{rsim_output}{An object of class Rsim}

\item{spname}{A character vector of species names to plot. Default is "all", which plots all species.}

\item{indplot}{A logical value indicating whether to plot individual species (TRUE) or all species (FALSE). Default is FALSE.}

\item{palette}{A character vector of colors to use for the plot. Default is "rainbow".}
}
\value{
An interactive plotly object.
}
\description{
Interactive plots for Rsim plot using plotly:: package and tidyverse::

Add another paragraph
}
\section{Contributors}{

ORIGINAL AUTHORS: rsim.plot function from Rpath by Kerim Aydin
}

\examples{
\dontrun{
# create the Rsim object
library(Rpath)
runObj <- rsim.scenario(rpath(REco.params),REco.params,years = 1:50)
Rsim.output <- rsim.run(runObj, years = 1:50)
# plot the Rsim object
rsim.plot.interactive(Rsim.output)
}



}
