% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsimplot_rpathviz.R
\name{rsimplot_rpathviz}
\alias{rsimplot_rpathviz}
\title{Interactive plots for Rsim}
\usage{
rsimplot_rpathviz(
  Rsim.output,
  spname = "all",
  indplot = FALSE,
  palette = "rainbow"
)
}
\arguments{
\item{Rsim.output}{An object of class Rsim}

\item{spname}{A character vector of species names to plot}

\item{indplot}{Logical value of whether to plot a single group or multiple groups}

\item{palette}{A character vector of colors to use}
}
\value{
An interactive plotly object
}
\description{
Interactive plots for Rsim plot using plotly:: package and tidyverse::

add another paragraph
}
\section{Contributors}{

rsim.plot function from Rpath by Kerim Aydin
}

\examples{
\dontrun{
# create the Rsim object
library(Rpath)
runObj <- rsim.scenario(rpath(REco.params), REco.params, years= 1:50)
Rsim.output <- rsim.run(runObj, years= 1:50)
# plot the Rsim object
rsimplot_rpathviz(Rsim.output)

}

}
