% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsimplot_rpathviz.R
\name{rsimplotviz}
\alias{rsimplotviz}
\title{Interactive plots for Rsim}
\usage{
rsimplotviz(
  Rsim.output,
  eco.name = NULL,
  spname = "all",
  rel_bio = FALSE,
  palette = "rsim_pal_dark"
)
}
\arguments{
\item{Rsim.output}{An object of class Rsim.}

\item{eco.name}{Optional name of the ecosystem.}

\item{spname}{A character vector of species names to plot. If "all", all species will be plotted and rel_bio will use the default FALSE.}

\item{rel_bio}{Logical value of whether to plot a single group biomass as relative or absolute value.}

\item{palette}{A character vector of colors to use.}
}
\value{
An interactive plotly object
}
\description{
Interactive plots for Rsim plot using plotly:: package and tidyverse::

add another paragraph
}
\section{Contributors}{

rsim.plot function from Rpath by Kerim Aydin
}

\examples{
\dontrun{
# create the Rsim object
runObj <- Rpath::rsim.scenario(Rpath::rpath(Rpath::REco.params), Rpath::REco.params, years= 1:50)
Rsim.output <- Rpath::rsim.run(runObj, years= 1:50)

# plot the Rsim object two default color blind friendly palettes
rsimplotviz(Rsim.output, eco.name= "Anchovy Bay", spname = "all",
palette= "rsim_pal_dark")
rsimplotviz(Rsim.output, eco.name= "Anchovy Bay", spname = "all",
palette= "rsim_pal_light")
rsimplotviz(Rsim.output, eco.name= "Anchovy Bay", spname = "Foragefish2",
rel_bio=FALSE, palette= "rsim_pal_light")

}

}
