% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webplot_rpathviz.R
\name{webplot_rpathviz}
\alias{webplot_rpathviz}
\title{Interactive plots for Rsim}
\usage{
webplot_rpathviz(
  Rpath.obj,
  eco.name = attr(Rpath.obj, "eco.name"),
  line.col = "grey",
  h_spacing = 3,
  fleet_color = "#B40F20",
  text_size = 3
)
}
\arguments{
\item{Rpath.obj}{An object of class Rpath \code{rpath()}.}

\item{eco.name}{Optional name of the ecosystem.  Default is the \code{eco.name} attribute from the
rpath object created from running \code{rpath()}.}

\item{line.col}{description.}

\item{h_spacing}{horizontal spacing multiplier, it spreads nodes horizontally.}

\item{fleet_color}{single color for fleet nodes.}

\item{text_size}{size of the text labels.}
}
\value{
Returns a plot visualization of the food web.
}
\description{
Interactive plots for Rsim plot using plotly:: package and tidyverse::

add another paragraph
}
\section{Contributors}{

webplot function from Rpath by Kerim Aydin
}

\examples{
\dontrun{
# Read in Rpath parameter file, generate and name model object
Rpath.obj <- Rpath::rpath(Rpath::AB.params, eco.name = "Anchovy Bay")
# Plot food web diagram with all groups labeled, including fleets
webplot_rpathviz(Rpath.obj, h_spacing = 3, text_size = 3)

}

}
